py/apps/mono-http/pyproject.toml
[tool.poetry]
name = "mono-http"
version = "0.1.0"
description = ""
authors = ["Mike Donnici <michael.donnici@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.0"
grpcio = "^1.43.0"

attributev1 = { path = "../../pkg/attributev1" }
statusv1 = { path = "../../pkg/statusv1" }
protobuf = "^3.19.4"


[tool.poetry.dev-dependencies]
pytest = "^6.2"
